signture cleanup of MVC functions

+++ src/api/routes/people/people-groups.php
@@ -19,7 +19,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
     });
 
     // get the group for the calendar, it's planned to only have the personan calendar and the calendar groups the user belongs to
-    $group->get('/calendars', function (Request $request, Response $response, array $args) {
+    $group->get('/calendars', function (Request $request, Response $response, array $args): Response
+    {
         $groups = GroupQuery::Create()
             ->orderByName()
             ->find();
@@ -36,7 +37,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderJSON($response, $return);
     });
 
-    $group->get('/groupsInCart', function (Request $request, Response $response, array $args) {
+    $group->get('/groupsInCart', function (Request $request, Response $response, array $args): Response
+    {
         $groupsInCart = [];
         $groups = GroupQuery::create()->find();
         foreach ($groups as $group) {
@@ -47,15 +49,18 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderJSON($response, ['groupsInCart' => $groupsInCart]);
     });
 
-    $group->get('/{groupID:[0-9]+}', function (Request $request, Response $response, array $args) {
+    $group->get('/{groupID:[0-9]+}', function (Request $request, Response $response, array $args): Response
+    {
         return SlimUtils::renderJSON($response, GroupQuery::create()->findOneById($args['groupID'])->toArray());
     });
 
-    $group->get('/{groupID:[0-9]+}/cartStatus', function (Request $request, Response $response, array $args) {
+    $group->get('/{groupID:[0-9]+}/cartStatus', function (Request $request, Response $response, array $args): Response
+    {
         return SlimUtils::renderJSON($response, GroupQuery::create()->findOneById($args['groupID'])->checkAgainstCart());
     });
 
-    $group->get('/{groupID:[0-9]+}/members', function (Request $request, Response $response, array $args) {
+    $group->get('/{groupID:[0-9]+}/members', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $members = Person2group2roleP2g2rQuery::create()
             ->joinWithPerson()
@@ -80,7 +85,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderJSON($response, $members->toArray());
     });
 
-    $group->get('/{groupID:[0-9]+}/events', function (Request $request, Response $response, array $args) {
+    $group->get('/{groupID:[0-9]+}/events', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $members = Person2group2roleP2g2rQuery::create()
             ->joinWithPerson()
@@ -88,7 +94,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderJSON($response, $members->toArray());
     });
 
-    $group->get('/{groupID:[0-9]+}/roles', function (Request $request, Response $response, array $args) {
+    $group->get('/{groupID:[0-9]+}/roles', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $group = GroupQuery::create()->findOneById($groupID);
         $roles = ListOptionQuery::create()->filterById($group->getRoleListId())->find();
@@ -97,7 +104,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
 });
 
 $app->group('/groups', function (RouteCollectorProxy $group) {
-    $group->post('/', function (Request $request, Response $response, array $args) {
+    $group->post('/', function (Request $request, Response $response, array $args): Response
+    {
         $groupSettings = $request->getParsedBody();
         $group = new Group();
         if ($groupSettings['isSundaySchool']) {
@@ -108,7 +116,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderJSON($response, $group->toArray());
     });
 
-    $group->post('/{groupID:[0-9]+}', function (Request $request, Response $response, array $args) {
+    $group->post('/{groupID:[0-9]+}', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $input = $request->getParsedBody();
         $group = GroupQuery::create()->findOneById($groupID);
@@ -119,13 +128,15 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderJSON($response, $group->toArray());
     });
 
-    $group->delete('/{groupID:[0-9]+}', function (Request $request, Response $response, array $args) {
+    $group->delete('/{groupID:[0-9]+}', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         GroupQuery::create()->findOneById($groupID)->delete();
         return SlimUtils::renderSuccessJSON($response);
     });
 
-    $group->delete('/{groupID:[0-9]+}/removeperson/{userID:[0-9]+}', function (Request $request, Response $response, array $args) {
+    $group->delete('/{groupID:[0-9]+}/removeperson/{userID:[0-9]+}', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $userID = $args['userID'];
         $person = PersonQuery::create()->findPk($userID);
@@ -145,7 +156,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderSuccessJSON($response);
     });
 
-    $group->post('/{groupID:[0-9]+}/addperson/{userID:[0-9]+}', function (Request $request, Response $response, array $args) {
+    $group->post('/{groupID:[0-9]+}/addperson/{userID:[0-9]+}', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $userID = $args['userID'];
         $person = PersonQuery::create()->findPk($userID);
@@ -176,7 +188,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderJSON($response, $members->toArray());
     });
 
-    $group->post('/{groupID:[0-9]+}/userRole/{userID:[0-9]+}', function (Request $request, Response $response, array $args) {
+    $group->post('/{groupID:[0-9]+}/userRole/{userID:[0-9]+}', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $userID = $args['userID'];
         $roleID = $request->getParsedBody()['roleID'];
@@ -186,7 +199,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderJSON($response, $membership->toArray());
     });
 
-    $group->post('/{groupID:[0-9]+}/roles/{roleID:[0-9]+}', function (Request $request, Response $response, array $args) {
+    $group->post('/{groupID:[0-9]+}/roles/{roleID:[0-9]+}', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $roleID = $args['roleID'];
         $input = $request->getParsedBody();
@@ -208,7 +222,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderJSON($response, ['success' => false]);
     });
 
-    $group->delete('/{groupID:[0-9]+}/roles/{roleID:[0-9]+}', function (Request $request, Response $response, array $args) {
+    $group->delete('/{groupID:[0-9]+}/roles/{roleID:[0-9]+}', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $roleID = $args['roleID'];
         $groupService = $this->get('GroupService');
@@ -216,7 +231,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderJSON($response, $groupService->deleteGroupRole($groupID, $roleID));
     });
 
-    $group->post('/{groupID:[0-9]+}/roles', function (Request $request, Response $response, array $args) {
+    $group->post('/{groupID:[0-9]+}/roles', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $roleName = $request->getParsedBody()['roleName'];
         $groupService = $this->get('GroupService');
@@ -224,7 +240,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderJSON($response, $groupService->addGroupRole($groupID, $roleName));
     });
 
-    $group->post('/{groupID:[0-9]+}/defaultRole', function (Request $request, Response $response, array $args) {
+    $group->post('/{groupID:[0-9]+}/defaultRole', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $roleID = $request->getParsedBody()['roleID'];
         $group = GroupQuery::create()->findPk($groupID);
@@ -233,7 +250,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         return SlimUtils::renderSuccessJSON($response);
     });
 
-    $group->post('/{groupID:[0-9]+}/setGroupSpecificPropertyStatus', function (Request $request, Response $response, array $args) {
+    $group->post('/{groupID:[0-9]+}/setGroupSpecificPropertyStatus', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $input = $request->getParsedBody();
         $groupService = $this->get('GroupService');
@@ -247,7 +265,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         }
     });
 
-    $group->post('/{groupID:[0-9]+}/settings/active/{value}', function (Request $request, Response $response, array $args) {
+    $group->post('/{groupID:[0-9]+}/settings/active/{value}', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $flag = $args['value'];
         if ($flag == 'true' || $flag == 'false') {
@@ -265,7 +284,8 @@ $app->group('/groups', function (RouteCollectorProxy $group) {
         }
     });
 
-    $group->post('/{groupID:[0-9]+}/settings/email/export/{value}', function (Request $request, Response $response, array $args) {
+    $group->post('/{groupID:[0-9]+}/settings/email/export/{value}', function (Request $request, Response $response, array $args): Response
+    {
         $groupID = $args['groupID'];
         $flag = $args['value'];
         if ($flag == 'true' || $flag == 'false') {
